Modify Containers and ContentObjects to use paths.
This replaces the use of IO streams.  The issue
with IO streams is that any attempted to do some
multiprocessing would be thwarted because file
handles can't be shared across processes.  (At
least not simply).

I see no real performance degredation when using
them in a single process as one is often likely to do.