Add NWN Objects.
These are class abstractions over GFF structures.  They are
currently read-only, so no values can be changed.  Instance
subclass are also not complete.  They are useable in so far
as the fields in the instances are identical to those in the
blueprint.  There are somecases where results would be wrong
like the item list of a creature/placeable/store instance
would contain RepositoryItem which is a pointer to a blueprint
and not an instance in itself.

TODO:
- Complete instances.
- Add value updating.